Changed:
* blue bottom line on menu bar - Done
* Rename Other competitive exam -> Other Exam - Done

*
Add home menu before register - Done
Remove date menu - Done
on click on home menu should go to the top - Done

* change in fetaure images - Done

*Change the alternate colors with the above section i.e. accordian - Done

* Remove black color from the footer - Done

* New section into the footer - Done

* Change the designs of the Privacy policy and Terms and conditions pages

*
Add quick links in footer section: - Done

Terms And Conditions - Done
Under no circumstances shall BRACE Education Academy be liable for any direct, indirect, special, incidental or consequential damages, including, but not limited to, loss of data or profit, arising out of the use, or the inability to use, the materials on this site, even if BRACE Education Academy or an authorized representative has been advised of the possibility of such damages. If your use of materials from this site results in the need for servicing, repair or correction of equipment or data, you assume any costs thereof should only be provided by the user of the application.

Privacy policy with email data - Done

* Why Brace (BEA) -> Why Brace Education Academy - Done
Also add spacing in bullet marks

*Remove address from contact form and add into the footer design is next page - Done